% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation}
\alias{annotation}
\title{Example Annotation}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 31 rows and 5 columns.
}
\usage{
data(annotation)
}
\description{
Example annotation of the gene sequences used in the example data sets
}
\keyword{datasets}
