% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_sfs}
\alias{calculate_sfs}
\title{Calculate SFS}
\usage{
calculate_sfs(alignment)
}
\arguments{
\item{alignment}{path to the alignment file or alignment imported as a DNAStringSet object}
}
\value{
a numeric vector of site frequency spectrum
}
\description{
This function calculates site frequency spectrum given a multiple sequnce alignment
}
\examples{
fasta<-paste0(path.package("gSoup"), "/alignment.fasta")
sfs<-calculate_sfs(fasta)


}
\author{
Piyal Karunarathne
}
