% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{div.stats}
\alias{div.stats}
\title{Diversity statistics}
\usage{
div.stats(alignment, pairwise_deletion = TRUE, pi_AS = FALSE)
}
\arguments{
\item{alignment}{multiple sequence alignment file path or
a DNAStringSet object}

\item{pairwise_deletion}{logical. whether to remove gaps in pairwise alignments}

\item{pi_AS}{logical.whether to calculate synonymous and non-synonymous polymorphism.
if \code{TRUE}, the alignment is considered to be only coding sequence}
}
\description{
This function calculates genetic diversity statistics
(e.g., \eqn{\pi, \theta, \pi[A], \pi[S]}, and Tajima's D)
for a complete sequence alignment. If you need to calculate these statistics
with a sliding window use the respective functions. e.g., \link{calculate_pi_theta}, \link{calculate_piAS}
}
