% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{fast_split_sw}
\alias{fast_split_sw}
\title{Sliding window split a multiple sequence alignment}
\usage{
fast_split_sw(
  fast_path,
  window = 500,
  steps = 100,
  export.fasta = TRUE,
  outpath = getwd()
)
}
\arguments{
\item{fast_path}{character. Path to the multiple sequnce alignment file}

\item{window}{numeric. Sliding window size}

\item{steps}{numeric. Step size}

\item{export.fasta}{logical. Whether to export the chunks in fasta format}

\item{outpath}{dir path. path to save the fasta files. If not provided,
fasta files will be saved to the working directory}
}
\value{
Returns chunks of multiple sequnce alignment split on a sliding window basis
}
\description{
This function splits a multiple sequence alignment with a sliding window and saves
chunks in fasta format
}
\author{
Piyal Karunarathne
}
