% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{find_blast_clusters}
\alias{find_blast_clusters}
\title{Find Blast Clusters
Group blast hits for sequence search based on a predefined proximity}
\usage{
find_blast_clusters(data, max_range = 4000)
}
\arguments{
\item{data}{output of blast search with columns \code{sstart} and \code{send}}

\item{max_range}{maximum expected length of the gene/sequence}
}
\description{
Find Blast Clusters
Group blast hits for sequence search based on a predefined proximity
}
