% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{gb2sense}
\alias{gb2sense}
\title{Function to extract DNA sequences and annotation information from .gb or .gbk files}
\usage{
gb2sense(
  gb_file,
  path = NULL,
  annotation = TRUE,
  fasta_out = NULL,
  annotation_out = NULL,
  separate_fasta = FALSE,
  DNAstringSet = TRUE,
  xy_orient = FALSE
)
}
\arguments{
\item{gb_file}{description}

\item{path}{description}

\item{annotation}{description}

\item{fasta_out}{description}

\item{annotation_out}{description}

\item{separate_fasta}{description}

\item{DNAstringSet}{description}

\item{xy_orient}{description}
}
\value{
Returns a DNAStringSet object of fasta sequnces and data frame of annotations
}
\description{
This function extracts the DNA sequence and annotation information from GeneBank (.gb/.gbk) files
and stores them in separate files
}
\examples{
\dontrun{
gb_out<-gb2sense(path = "./gb_files/",DNAstringSet = TRUE,separate_fasta = TRUE)
gb_out<-gb2sense(gb_file = "./gb_files/Lan5_HMA4-1.gb",DNAstringSet = TRUE,separate_fasta = TRUE)}

}
\author{
Piyal Karunarathne
}
