% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{GeneAnno}
\alias{GeneAnno}
\title{Annotate gene structure
This function annotates gene structure predictions based on .gff or other annotation outputs}
\usage{
GeneAnno(
  an.tab,
  genes = NULL,
  scale = c("mb", "kb"),
  orient = c("horizontal", "vertical"),
  ann.width = 0.4,
  up_down_stream = FALSE,
  gname.col = "seq",
  intron = FALSE,
  draw.axis = TRUE,
  label = TRUE,
  justified = TRUE,
  ...
)
}
\arguments{
\item{an.tab}{description}

\item{genes}{description}

\item{scale}{description}

\item{orient}{description}

\item{ann.width}{description}

\item{up_down_stream}{description}

\item{gname.col}{description}

\item{intron}{desc}

\item{draw.axis}{description}

\item{label}{description}

\item{justified}{description}

\item{...}{description}
}
\description{
Annotate gene structure
This function annotates gene structure predictions based on .gff or other annotation outputs
}
\author{
Piyal Karunarathne
}
