% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{MKT_HKA_test}
\alias{MKT_HKA_test}
\title{MKT and HKA test statistics}
\usage{
MKT_HKA_test(dna_algn, outgroup_name, region = NULL, detailed_output = FALSE)
}
\arguments{
\item{dna_algn}{DNA sequence alignment in DNAStringSet object or path to the alignment file in FASTA format}

\item{outgroup_name}{Character. Name of the outgroup sequences exactly as in the sequnece alignment}

\item{region}{Numeric. Range of the alignment if the complete alignment should not be considered for the
test (e.g. range=c(200-2000))}

\item{detailed_output}{Logical. If \code{TRUE} a detailed output will be printed}
}
\value{
Returns a list with test statistics
}
\description{
This function calculates McDonald Kreitman Test (MKT) and Hudson-Kreitman-Aguadé (HKA) Test
and outputs variable sites and test significance
}
\author{
Piyal Karunarathne
}
