% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{overlap_percentage}
\alias{overlap_percentage}
\title{Overlapping percentage of two blast hits}
\usage{
overlap_percentage(start1, end1, start2, end2)
}
\arguments{
\item{start1}{starting position of first hit}

\item{end1}{end position of first hit}

\item{start2}{starting position of second hit}

\item{end2}{end position of second hit}
}
\description{
This function will report the overlapping percentage of two blast hits.
It is useful in finding removing overlapping hits, especially when multicopy
regions return hits for the same gene region or wise-versa.
}
\details{
This function is specifically useful if you have search for multiple copies of genes
where same region of a genome returs hits for different copies. You can run this of a table
to find overlapping hits for multiple BLAST searches.
}
\examples{
\dontrun{for (j in (i + 1):nrow(df))
 {if (overlap_percentage(df$start[i], df$end[i], df$start[j], df$end[j]) > 0.7)
  {overlapping_rows<-c(unique(overlapping_rows),j)}}}

}
\author{
Piyal Karunarathne
}
