% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{pairwise_length}
\alias{pairwise_length}
\title{Determine actual pairwise alignment length}
\usage{
pairwise_length(msa_path)
}
\arguments{
\item{msa_path}{character. path to the multiple sequnce alignment in fasta format}
}
\value{
A matrix of pairwise true alignment length without gaps
}
\description{
This function removes gaps from a multiple sequence alignment
with a pairwise assessment and returns a matrix with pairwise alignment
length
}
\examples{
fasta<-paste0(path.package("gSoup"),"/alignment.fasta")
no_gap_lengths<-pairwise_length(fasta)

}
\author{
Piyal Karunarathne
}
