% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pi_and_theta}
\alias{pi_and_theta}
\title{Example pi and theta ouput}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(pi_and_theta)
}
\description{
Example pi and theta ouput
}
\keyword{datasets}
