% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pi_theta}
\alias{pi_theta}
\title{Calculate nucleotide diversity \eqn{\theta[\pi]}}
\usage{
pi_theta(
  fast_path,
  window = 100,
  pairwise_deletion = TRUE,
  plot = TRUE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{fast_path}{description}

\item{window}{description}

\item{pairwise_deletion}{description}

\item{plot}{description}

\item{progress}{description}

\item{...}{description}
}
\description{
This function calculates nucleotide diversity and Watterson's theta
from a multiple sequnce alignment
}
\author{
Piyal Karunarathne
}
