% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{piA_and_S}
\alias{piA_and_S}
\title{Example \eqn{\pi[A] \pi[S]} dataset}
\format{
An object of class \code{list} of length 5.
}
\usage{
data(piA_and_S)
}
\description{
Example \eqn{\pi[A] \pi[S]} dataset
}
\keyword{datasets}
