% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{piAS}
\alias{piAS}
\title{Synonymous and non-synonymous polymorphism}
\usage{
piAS(
  algn,
  annotation,
  window = 100,
  gene_length = NULL,
  pairwise_deletion = TRUE
)
}
\arguments{
\item{algn}{description}

\item{annotation}{description}

\item{window}{description}

\item{gene_length}{description}

\item{pairwise_deletion}{description}
}
\description{
This function calculates sysnonymous and non-synonymous polymorphism
from multiple sequnce alignment of coding regions of a gene
}
\author{
Piyal Karunarathne
}
