% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_pi_theta}
\alias{plot_pi_theta}
\title{Plot nucleotide diversity}
\usage{
plot_pi_theta(
  Pi,
  annotation,
  anno.type = c("exon", "CDS"),
  ratio = FALSE,
  log_scale = TRUE,
  ann.width = NULL,
  mid.pos = NULL,
  intron = FALSE,
  ...
)
}
\arguments{
\item{Pi}{description}

\item{annotation}{description}

\item{anno.type}{description}

\item{ratio}{description}

\item{log_scale}{description}

\item{ann.width}{description}

\item{mid.pos}{description}

\item{intron}{description}

\item{...}{description}
}
\description{
This function plots nucleotide diversity \eqn{\theta[\pi]} and \eqn{\theta[\W]}
calculated with \link{pi_theta} function and plots it above the annotation of
the gene.
}
\author{
Piyal Karunarathne
}
