% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_piAS}
\alias{plot_piAS}
\title{Plot synonymous non-synonymous polymorphism}
\usage{
plot_piAS(
  pis,
  annotation,
  anno.type = c("CDS", "exon"),
  ratio = FALSE,
  log_scale = TRUE,
  ann.width = NULL,
  mid.pos = NULL,
  ...
)
}
\arguments{
\item{pis}{description}

\item{annotation}{description}

\item{anno.type}{description}

\item{ratio}{description}

\item{log_scale}{description}

\item{ann.width}{description}

\item{mid.pos}{description}

\item{...}{description}
}
\description{
This function plots synonymous \eqn{\pi[S]} non-synonymous \eqn{\pi[A]} polymorphism
and their ratio \eqn{\pi[A]/\pi[S]} calculated with \link{piAS} function
and plots it above the annotation of the gene.
}
\author{
Piyal Karunarathne
}
