% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{win_extrapol}
\alias{win_extrapol}
\title{Map window extrapolation}
\usage{
win_extrapol(spat, window = 50, width = 3, show_progress = TRUE)
}
\arguments{
\item{spat}{An object of SpatRaster or a matrix created from a SpatRaster}

\item{window}{Numeric. size of the extrapolation window}

\item{width}{Numeric. number of cells used for calculating the average}

\item{show_progress}{Logical. Whether to show the progress bar}
}
\value{
Returns an extrapolated matrix
}
\description{
This function extrapolates maps using a given window size to given extents
}
\author{
Piyal Karunarathne
}
