% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{gb2sense}
\alias{gb2sense}
\title{Function to extract DNA sequences and annotation information from .gb or .gbk files}
\usage{
gb2sense(
  gb_file,
  path = NULL,
  annotation = TRUE,
  fasta_out = NULL,
  annotation_out = NULL,
  separate_fasta = FALSE,
  DNAstringSet = TRUE,
  LR_orient = FALSE
)
}
\arguments{
\item{gb_file}{path to the .gb(k) file}

\item{path}{path the genebank file directory with multipel .gb(k) files}

\item{annotation}{logical. Whether to extract annotation information}

\item{fasta_out}{path to output fasta file}

\item{annotation_out}{path to output the annotation file}

\item{separate_fasta}{logical. Whether to export separate fasta files for multiple .gb(k) files}

\item{DNAstringSet}{logical. Whether to return the fasta sequnces in a DNAStringSet object}

\item{LR_orient}{logical. Whether to reorient the annotation when the sequence is on the negative strand}
}
\value{
Returns a DNAStringSet object of fasta sequnces and data frame of annotations
}
\description{
This function extracts the DNA sequence and annotation information from GeneBank (.gb/.gbk) files
and stores them in separate files
}
\examples{
\dontrun{
gb_out<-gb2sense(path = "./gb_files/",DNAstringSet = TRUE,separate_fasta = TRUE)
gb_out<-gb2sense(gb_file = "./gb_files/Lan5_HMA4-1.gb",DNAstringSet = TRUE,separate_fasta = TRUE)}

}
\author{
Piyal Karunarathne
}
