% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{pairwise_length}
\alias{pairwise_length}
\title{Remove all gaps from a pairwise sequence alignment}
\usage{
pairwise_length(msa_path)
}
\arguments{
\item{msa_path}{character. path to the multiple sequnce alignment in fasta format}
}
\value{
A matrix of pairwise true alignment length without gaps
}
\description{
This function removes gaps from a multiple sequence alignment
with a pairwise assessment and returns a matrix with pairwise alignment
length
}
\author{
Piyal Karunarathne
}
