% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_piAS}
\alias{plot_piAS}
\title{Plot synonymous non-synonymous polymorphism}
\usage{
plot_piAS(
  pis,
  annotation,
  anno.type = c("CDS", "exon"),
  ratio = FALSE,
  log_scale = TRUE,
  ann.width = NULL,
  mid.pos = NULL,
  window_size = 2,
  ...
)
}
\arguments{
\item{pis}{list containing \eqn{\pi[A/S]} from the output of \link{calculate_piAS} function}

\item{annotation}{a data frame of the annotation of the gene or the sequence in the alignment.
must have at least the three colums \code{type, start, end}}

\item{anno.type}{character. type of gene structure to be plotted (e.g. \code{exon} or \code{CDS})}

\item{ratio}{logical. Whether to draw the \eqn{\pi[A]/\pi[S]} ratio plot}

\item{log_scale}{logical. If \code{ratio=TRUE}, whether to use the log2 scale}

\item{ann.width}{numerical. width of the gene annotation as a percentage at the bottom of the plot. \code{default=NULL}, determined as a percentage of the plot size}

\item{mid.pos}{numeric. mid position of the annotation plot in respect to the y axis}

\item{window_size}{numeric. Window size for smoothing the curve}

\item{...}{other parameters to be passed to link{plot.default}}
}
\description{
This function plots synonymous \eqn{\pi[S]} non-synonymous \eqn{\pi[A]} polymorphism
and their ratio \eqn{\pi[A]/\pi[S]} calculated with \link{calculate_piAS} function
and plots it above the annotation of the gene.
}
\examples{
data(annotation)
data(piA_and_S)
plot_piAS(piA_and_S,annotation)


}
\author{
Piyal Karunarathne
}
