% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_proc.R
\name{rm_gaps}
\alias{rm_gaps}
\title{Remove gaps from a reference in an alignment}
\usage{
rm_gaps(fast_path, ref, export.fasta = TRUE)
}
\arguments{
\item{fast_path}{either the path to a msa .fasta file or alignment imported as DNAStringSet object}

\item{ref}{name of the reference sequence in the alignment}

\item{export.fasta}{logical/path. whether to export gaps removed alignment in fasta format.
If TRUE, it will be saved to the working directory, or to the path provided in \code{export.fasta}}
}
\description{
This function removes gaps from fasta multiple sequence alignment to have no gaps in the reference sequences
}
\author{
Piyal Karunarathne
}
